+++ CHANGELOG.md
@@ -1,6 +1,8 @@
  == CHANGELOG ==
 
-* Reworked archive loading, added new file menu option to browse an archive folder.
+#### NifSkope-2.0.dev9-20240428
+
+* Reworked archive loading, added new file menu option to browse a game or archive folder.
 * The archive browser now displays uncompressed file sizes from BA2 archives.
 * Added support for version 7 and 8 BA2 archives used by the Fallout 4 next gen update.
 * The glass settings added to version 21 Fallout 76 effect materials are now displayed under the material properties.